#24 renamed graph_view.py
