+++ src/barplot_graphics.py
